Cast torch input tensors to dtype that is supported in forge.  (#1335)
Remove manual casting in tests.

### Ticket
Issue #952 

### Problem description
There was a bug when input tensors have dtype that is not supported in
Forge (for example int64). Compiler handled inputs like they were some
lower dtype, while actual runtime tensors stayed in original dtype. This
is explained in detail in issue #952

### What's changed
Now we cast torch tensor to appropriate dtype based on forge support.

+++ forge/test/mlir/operators/nn/test_nn.py
@@ -297,7 +297,7 @@ def test_embedding(vocab_size, token_num, embedding_dim):
             return self.embedding(x)
 
     inputs = [
-        torch.randint(0, vocab_size, (1, token_num)).to(torch.int32),
+        torch.randint(0, vocab_size, (1, token_num)),
     ]
 
     framework_model = Embedding()
